pyproject.toml
[project]
name = "umrah-trip-creator"
version = "0.1.0"
description = "Multi-agent system for Umrah trip planning using Amazon Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0",
    "bedrock-agentcore-runtime>=0.1.0",
    "strands-agents>=0.1.0",
    "anthropic>=0.40.0",
    "openai>=1.54.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
